+++ api/controllers/posts.js
@@ -49,26 +49,6 @@ const PostsController = {
       res.status(200).json({ posts: result, token: token });
     }
   }
-
-  // AddImage: async (req, res) => {
-  //   const filename = req.body.filename;
-  //   const post_id = req.body.post_id;
-  //   console.log("BACKEND RECIEVED DATA: ", filename, post_id);
-
-  //   // Check if the request contains filename and post id
-  //   if (!filename || !post_id) {
-  //     console.log()
-  //     return res.status(400).json({ message: "Bad request" });
-  //   }
-
-  //   // replace filename: null with new filename
-  //   const result = await Post.findOneAndUpdate(
-  //     { _id: post_id },
-  //     { $set: { image_path: filename } },
-  //   );
-
-  //   res.status(200).json({ message: "GOT FILENAME - OK" });
-  // },
 };
 
 module.exports = PostsController;
