+++ api/package.json
@@ -17,6 +17,7 @@
     "debug": "~2.6.9",
     "express": "^4.18.2",
     "express-session": "^1.17.2",
+    "formdata-node": "^6.0.3",
     "hbs": "^4.1.0",
     "http-errors": "~1.6.3",
     "jsonwebtoken": "^9.0.0",
@@ -30,6 +31,7 @@
   },
   "devDependencies": {
     "jest": "^27.5.1",
+    "jest-fetch-mock": "^3.0.3",
     "mock-fs": "^5.2.0"
   },
   "nodemonConfig": {
