+++ package.json
@@ -4,6 +4,7 @@
     "eslint": "^8.53.0",
     "file": "^0.2.2",
     "form-data": "^4.0.0",
+    "formdata-node": "^6.0.3",
     "multer": "^1.4.5-lts.1"
   },
   "devDependencies": {
